+++ src/main/java/org/kie/processmigration/service/KieService.java
@@ -49,7 +49,7 @@ public interface KieService {
 
     List<RunningInstance> getRunningInstances(String kieServerId, String containerId, Integer page, Integer pageSize, String sortBy, String orderBy) throws InvalidKieServerException;
 
-    Integer countRunningInstances(String kieServerId, String containerId) throws InvalidKieServerException;
+    Long countRunningInstances(String kieServerId, String containerId) throws InvalidKieServerException;
 
     KieServicesClient getClient(String kieServerId) throws InvalidKieServerException;
 }
